removed hello folder and renamed some directories
