--- a1_hello/HelloWorld.java
@@ -1,11 +1,0 @@
-/**
- * @assignment: Write a program to print Hello World
- */
-
-package a1_hello;
-
-public class HelloWorld {
-    public static void main(String[] args) {
-        System.out.println("Hello World");
-    }
-}
