+++ a1_search_algo/Problems/p04_FindElementInPeakArray.java
@@ -1,4 +1,4 @@
-package a2_search_algo.Problems;
+package a1_search_algo.Problems;
 
 public class p04_FindElementInPeakArray {
