+++ a2_sorting_algo/BubbleSort.java
@@ -1,4 +1,4 @@
-package a3_sorting_algo;
+package a2_sorting_algo;
 
 public class BubbleSort {
     Package a3_sorting_algo;
