+++ a2_sorting_algo/InsersionSort.java
@@ -1,4 +1,4 @@
-package a3_sorting_algo;
+package a2_sorting_algo;
 
 public class InsersionSort {
     public static void main(String[] args) {
