+++ a2_sorting_algo/SelectionSort.java
@@ -1,4 +1,4 @@
-package a3_sorting_algo;
+package a2_sorting_algo;
 
 public class SelectionSort {
     public static void main(String[] args) {
